Abstract class name should be prefixed with "Abstract"
Keep and deprecate existing `FormattedTimestampJsonProvider` class until next major version
