+++ src/main/java/net/logstash/logback/composite/accessevent/AccessEventFormattedTimestampJsonProvider.java
@@ -15,12 +15,12 @@
  */
 package net.logstash.logback.composite.accessevent;
 
-import net.logstash.logback.composite.FormattedTimestampJsonProvider;
+import net.logstash.logback.composite.AbstractFormattedTimestampJsonProvider;
 import net.logstash.logback.fieldnames.LogstashAccessFieldNames;
 
 import ch.qos.logback.access.spi.IAccessEvent;
 
-public class AccessEventFormattedTimestampJsonProvider extends FormattedTimestampJsonProvider<IAccessEvent, LogstashAccessFieldNames> {
+public class AccessEventFormattedTimestampJsonProvider extends AbstractFormattedTimestampJsonProvider<IAccessEvent, LogstashAccessFieldNames> {
 
     @Override
     protected long getTimestampAsMillis(IAccessEvent event) {
